Migrating the setting up and tearing down modules
